.gitconfig|.ps1: updates for working with detached HEAD
I work in non-branches a lot, so the warning message about entering the
state is just annoying. Additionally, it's nice to know at a glance when
the current detached-head state is due to a tag checkout, so now .ps1
prints a <tag> instead of just the SHA prefix.

`git detach` checks out the current HEAD sha, making it easier to clean
up a branch after pushing to remote.